(PC-35679) refactor(offer): add OfferImageCarousel web file
